+++ core/src/test/java/org/svip/sbomfactory/serializers/serializer/SVIPSBOMJSONSerializerTest.java
@@ -1,0 +1,16 @@
+package org.svip.sbomfactory.serializers.serializer;
+
+import com.fasterxml.jackson.core.JsonProcessingException;
+import org.junit.jupiter.api.Test;
+import org.svip.sbomfactory.generators.utils.Debug;
+
+public class SVIPSBOMJSONSerializerTest extends SerializerTest {
+    public SVIPSBOMJSONSerializerTest() {
+        super(new SVIPSBOMJSONSerializer());
+    }
+
+    @Test
+    public void writeToStringTest() throws JsonProcessingException {
+        Debug.log(Debug.LOG_TYPE.SUMMARY, getSerializer().writeToString(getTestSBOM()));
+    }
+}